Move update preferred name to seperate file
